fashion-attribute/requirements.txt
numpy==2.3.3
pandas==2.3.2
pandas-stubs==2.3.2.250926
python-dateutil==2.9.0.post0
Pillow==11.3.0
transformers>=4.40.0
torch>=2.9.0
torchvision>=0.24.0
tqdm==4.67.1
sentence-transformers==5.1.1